Add files via upload
The LOOKUP, INDEX, MATCH, and SUMIFS functions are all powerful tools that can be used to extract data from Excel spreadsheets. They each have their own strengths and weaknesses, so it is important to choose the right function for the task at hand.

LOOKUP: The LOOKUP function is a simple function that can be used to find a value in a range of cells. It takes two arguments: the lookup value and the range of cells to search. The LOOKUP function will return the value in the range that matches the lookup value.
INDEX: The INDEX function is a more powerful function that can be used to extract data from a range of cells. It takes three arguments: the range of cells to search, the row number, and the column number. The INDEX function will return the value at the specified row and column in the range of cells.
MATCH: The MATCH function is a function that can be used to find the position of a value in a range of cells. It takes two arguments: the lookup value and the range of cells to search. The MATCH function will return the position of the lookup value in the range, starting from 1.
SUMIFS: The SUMIFS function is a function that can be used to sum the values in a range of cells that meet certain criteria. It takes three or more arguments: the range of cells to sum, the criteria ranges, and the criteria values. The SUMIFS function will sum the values in the range of cells that meet all the criteria.
In general, the LOOKUP function is the fastest function, but it is also the least flexible. The INDEX function is more flexible, but it is also slower. The MATCH function is the most flexible function, but it is also the slowest. The SUMIFS function is a good compromise between speed and flexibility.