Add files via upload
Excel's Scenario Manager and Solver are two powerful tools that can be used for data modeling and what-if analysis.

Scenario Manager allows you to create and save different sets of values for your worksheet, which you can then switch between to see how different values affect the results. This can be useful for comparing different scenarios, such as different sales forecasts or marketing plans.
Solver is a more advanced tool that allows you to find the optimal solution to a problem. For example, you could use Solver to find the sales price that would maximize your profits, or the marketing budget that would maximize your market share.
The main difference between Scenario Manager and Solver is that Scenario Manager is designed for simple what-if analysis, while Solver is designed for more complex problems. Scenario Manager is also easier to use, while Solver can be more powerful.